Update MSRV to 1.81
This also changes to use the derived `Default` for enums since
that is available with a later MSRV (and would otherwise have
clippy warnings).

Updating the MSRV will allow the `no_std` support to be slightly
better by allowing `core::error::Error` to be impl'd for the
error types rather than only having that for `std` builds.